SDN-1223 Add config knob to carry inner tos value to outer header
OVN supports indicating a ToS value in the outer geneve header.
It relies on OVS external_ids:ovn-encap-tos to indicate the value
(including inherit). OVN-K8s node can set this external id based on
configuration.

Currently, we set this only on the ovn-k8s node runinng on the
smart nic (dpu) since that is in the datpath for the applications
we are mainly concerned about (tenant traffic). We could add this
to the other nodes, e.g. control nodes, if needed.

Signed-off-by: venu iyer <venugopali@nvidia.com>