Changes for multicluster underlay network design
For NGN, we are using a multi-vlan underlay network design with
independent vlans for infra, tenant and overlay networks.
To support multiple ovnkube node instances, there are separate physnet
bridges for infra and tenant which hold their corresponding vlans.

The following changes are made to support this design.

* Support for configuring conntrack zone than the default (64000). Via
environmental variable OVN_CONNTRACK_ZONE. Improved support in the code
to receive and set this parameter correctly.

* Set Uplink port. It is now possible to set the port that is used as the
uplink port directly. This is useful in scenarios like above where the
gateway bridge is not directly connected to the uplink interface and we
cannot autodetect the uplink correctly. Flag: `--gateway-uplink-port` in
environment variable OVN_GATEWAY_OPTS.

* Set PhysicalNetwork Key name. The name of the key to use to find the
OVS bridge that provides access to a particular instance of ovnkube-node.
This is needed since we have multiple bridges as external access and
we want to select the bridge mapping using this key.
Flag: `--gateway-physnetname-key` in environment variable OVN_GATEWAY_OPTS.

In addition to that

* Make Loadbalancer configuration for services multicluster comptible.
   - Prefix with cluster_name, add cluster_name external_id
   - Add loadbalancer port filtering to prevent staleLB detection
 seeing LBs from another cluster.

* Support Pods with same name in multiple clusters

 Adds support for running pods with the name in multiple
 clusters. ovn-kubernetes was adding logical ports with the format
 `namespace_podname`. In the multicluster case with shared OVN,
 these could collide. To avoid this we add the cluster_name prefix to
 the logical port on the OVN side as well as the OVS side (via the CNI).

Signed-off-by: Hareesh Puthalath <hareeshp@nvidia.com>