[SDN-1471] SS VMIs doesn't always come up with all interfaces on storage
in UpdateFunc() we need to iterate through all the NADs that the Pod
connects to. we were returning after the 1st iteration.

the networkMap will have two entries -- services1 and services2. the
order in which we are going to iterate is not guaranteed. so, say we are
adding the 1st interface to the pod and that the Connection details is
set only for the 1st interface, however in the for loop in updateFunc()
if we lookup the connection details for the 2nd interface first, then
it will return nil and we will return from the UpdateFunc itself
instead of iterating to the next `nadName` in the Map.