duplicated IPs are allocated to multiple Pods
Note that retryPod could be added by addNode handler. To avoid duplicate
IPs being allocated to multiple pods, start retryPod handler and get
ready to handle the retryPod after the IPs of the existing Pods are
reserved in syncPods(). Only then those IPs won't be reallocated to any
new Pods (that are added when ovnkube-master is down/restarting).

Signed-off-by: Yun Zhou <yunz@nvidia.com>