SDN 1221 TCP DoS mitigation for ARM traffic on public interfaces
This also includes
SDN 1129 Configuring DoS service on ARM BF2 for GFN RP service

For networks that need to be protected using xdp/synproxy, we'll use an
indication - XDPService, in the network attachment definitions.

When a network attachment def with this annotation is created, we'll set
up the xdp namespace (information provided as config to ovn-k8s running
on the DPU), with the route table for that public network and some
additional information, including gw mac.

When a pod/vmi is created on that network, we'll add OF rules for that
pod/vmi's IP on the phys OVS bridge to configure xdp/synproxy in between
the pod/vmi and external. Rules are also added to bypass xdp for traffic
originating within the pod/VMI. Any TCP traffic originating from
external to this pod/VMI will be subject to synproxy using xdp.

ovn-k8s will healthcheck the localnet patch port and also the OF flows
added for the xdp backed networks. Currently, we don't healtcheck the
xdp ns itself.

Signed-off-by: venu iyer <venugopali@nvidia.com>