ovnkube-node: Full mode management port
Currently in full mode ovnkube uses OVS internal port as management
interface. However in some use cases it is preferable to use real
devices, like SRIOV VFs, as management ports to utilize offloading. To
do that configure both netdevice and its representor if
ovnkube-node-mgmt-port-netdev and ovnkube-node-mgmt-port-representor
parameters are specified. If parameters values are empty, OVS internal
port configured as it is now.
Since representor resides on the same host, it is renamed to
"ovn-k8s-mp0_0".

Allign existing and add new unit tests.

Signed-off-by: Dmytro Linkin <dlinkin@nvidia.com>