Do not remove ovn-kubernetes CNI conf file when running in DPU mode
Currently we need to run two instances of ovnk8s-node container in the DPU
a) full mode  -> for supporting SF pods in the DPU (Infra cluster)
b) dpu mode -> for supporting vf plugging on the x86 side. (Tenant cluster)

What happens is that ovnk8s-node in full mode creates the 10-ovn-kubernetes.conf file.
Then when as part of the Tenant cluster bringup, ovnk8s-node in dpu mode starts and it
deletes the 10-ovn-kubernetes.conf file.

Earlier this was a trap, but now its an explicit delete on startup.

Signed-off-by: Alin-Gabriel Serdean <aserdean@ovn.org>