vendor/libovsdb: Move NewMonitorRequest from mapper to client.
The function NewMonitorRequest was a member of the mapper type, but it
doesn't depend on mapper object any more. It is not closely related to
mapper, so move it to the client module, where it is used.

Signed-off-by: Han Zhou <hzhou@ovn.org>
(cherry picked from commit 9b84c5de1a354db990c1335474f0338e37fc7389)
Signed-off-by: Yun Zhou <yunz@nvidia.com>