do not fail if the same VF representor is added for the same Pod/NAD
In case when DPU reboots, ovnkube-node might miss the connectionDetails
annotation update events if there is any cmdDel/cmdAdd sequence
happening in between. As the result, when it adds the newly assigned VF
representor to br-int with the new sandbox, it fails and complaining
that VF representor is already added with different sandbox. This commit
is to fix that.

Note that if different VF representor interface is assigned in this
case, the old one will be deleted from br-int before the new one is
added when ConfigureOVS() finds they share the same iface-id.

Signed-off-by: Yun Zhou <yunz@nvidia.com>
(cherry picked from commit 8f5c90a48d4c85830e9437ee255a6f64faa803e4)
Signed-off-by: Yun Zhou <yunz@nvidia.com>