SDN-2068: Improve the ovnkube_node_cni_request_duration_seconds_bucket metrics
The buckets defined by the metrics was:
prometheus.ExponentialBuckets(.1, 2, 15)

which expands to:
[0.1, 0.2, 0.4, 0.8, 1.6, 3.2, 6.4, 12.8, 25.6, 51.2, 102.4, ...]

This doesn't provide accurate information for analysis because most of
our P99 latency falls in the range between 1.6 - 6.4, so we only see the
value 1.6, 3.2, 6.4, while it would be better to see if it is for
example, 2, 3, 4, or 5 seconds for the CNI request duration. This patch
changes it to use linearBuckets instead of Exponential to better fit our
use case.

Signed-off-by: Han Zhou <hzhou@ovn.org>
(cherry picked from commit 85fea825458d339184240ac147ae6d9bf2a9750c)
Signed-off-by: Yun Zhou <yunz@nvidia.com>