Improve syncNodes to remove stale data
Partial removal of nodes can leave OVN database with the gateway
logical router and external logical switch of the removed node.
This commit improves syncNode functionality, so these cases are
also taken into consideration.

When deleting a port from a logical switch or router, it is
possible that the logical switch or router themselves have been
deleted. This commit changes the behavior of delete so it
is not an error when if that happens.

Reported-at: https://issues.redhat.com/browse/OCPBUGS-13744
Signed-off-by: Flavio Fernandes <flaviof@redhat.com>
(cherry picked from commit 487bba32a5f6256e53b77f8f43808c74763507cf)
Signed-off-by: Yun Zhou <yunz@nvidia.com>