specify representor metering (mrl) label key in cli
ovnkube-node need x86 host's type to determine miss rate config if it
is specified in NAD. It was stored in k8s node's labels, and the label
key (`ngn2.nvidia.com/hosttype`) was hard coded. This commit makes the
label key a cli param, so the code won't look specific to NGN, and  users
may ignore it if it's not required for their use cases.

Signed-off-by: Xiaobin Qu <xqu@nvidia.com>