[SDN-2275] Ensure OVSDB RAFT schema upgrade only from leader node
In large scale zones, the schema conversion takes relatively long time (> 1s).
If the zone is very active, with frequent changes during the schema conversion,
and if the schema conversion is triggered from a follower node, then when the
change proposed to the leader node, it always end up with a transaction conflict
and retry. This would never be successful and in the end the schema conversion
command at the follower node would time out and fail.

To address this issue, we add a function in ovndb-raft-function to
ensure the schema upgrade from the leader pod.

Signed-off-by: Zhen Wang <zhewang@nvidia.com>