SDN-2361: Set CIDR based MAC binding aging threshold for cluster subnets.
It is observed that in large deployments, there are occasional MAC
binding entries populated to SB during pod deletions, even though the
pod's LSP had static IP and MAC configured and should never require
dynamic MAC binding population. Over time, huge amount of MAC-binding
entries accumulates, resulting in huge amount of MAC-binding flows on
every chassis.

It happened because when the pod is being deleted, the related lflow
deletions from SB are sent to different chassis with different latency.
If there are other pods keep sending traffic to the pod being deleted,
and if the source chassis's static ARP resolve flows are deleted before
the dest chassis's ARP responder flows, ARP request would be triggered
from the source side trying to dynamically resolve the dst IP, and the
dest chassis's LS ARP responder would reply with ARP response, which in
turn would trigger the MAC binding learning when the response hits the
LR pipeline.

Originally, we don't enable MAC binding aging, because the addresses of
LSPs were not expected to be populated to MAC binding table. However,
for this problem, it is hard to avoid the SB update race completely. So
enabling aging becomes a reasonable approach to cleaning up these
unnecessary MAC bindings, while we don't want to enable aging for other
valid entries (we have use cases in the ovn-k8s cluster router that
depend on dynamic MAC binding). So with this patch, we can disable aging
by default, but enable it for the pod IP space. The threshold is default
to 36000 (10 hours).

Signed-off-by: Han Zhou <hzhou@ovn.org>