SDN-4111: Add DAN support for Kata Containers with VFIO interfaces
When using Kata Container with a VFIO interface, the network needs to be
configured via DAN. This change adds DAN support in ovn-k8s CNI.
When the network interface is VFIO, it generates the DAN config file for
the sandbox in the folder specified by the CNI NetConf option
`kata_dan_conf_dir`.

Hardcode kata_dan_conf_dir to a fixed path and introduce
OVNKUBE_ENABLE_KATA_DAN as the feature flag.

Signed-off-by: Lei Huang <leih@nvidia.com>