[SDN-2814] ovnkube.sh: use underscore instead of dash to avoid execution
This is a follow up from commit 7d188a2c8a826020413d06270b5b850c95bf6366
This addresses the following issue when executing the script:

    /root/ovnkube.sh: line 393: ovnkube-wait-on-ovn-install-extid=false: command not found

In Bash, the hyphen (-) is not a valid character for variable names.
Therefore, when Bash encounters this line, it interprets
ovnkube-wait-on-ovn-install-extid=false as an attempt to execute a
command named ovnkube-wait-on-ovn-install-extid with an argument =false.
To fix this, use underscores (_) instead of hyphens (-) in variable
name.

Signed-off-by: Flavio Fernandes <ffernandes@nvidia.com>
(cherry picked from commit e3f95f925294d4b8db80f995cd571f1a09ac6b10)
Signed-off-by: Yun Zhou <yunz@nvidia.com>