Strip vlan before outputting to hostports.
When gateway is configured with a VLAN ID, the packets
coming from physical interface for VIPs and externalIPs
have vlan tags in them. We need to strip that vlan
before sending the packets to the access ports.

Note for upstream: Currently VLAN Gateways are not upstreamed
by Nvidia. When they are upstreamed, we should upstream this too.

Signed-off-by: Guru Shetty <gurus@nvidia.com>
(cherry picked from commit 099db098a3fdd8ae42992df63f773cf366569a77)
Signed-off-by: Yun Zhou <yunz@nvidia.com>