SDN-2614 Public LS traffic not offloaded to HW
The issue is that all the packets from the wire is always forwarded to
ct_zone=64000 and then to table=1 to see if the packets belong to OVN
or host. So, the reply packets for the Pod's request packets on the
Pod's public interface is not offloaded at all.

The fix here is to forward the packets to table=1 only if the dst
MAC address matches the shared (between host and OVN) MAC address.

Signed-off-by: Girish Moodalbail <gmoodalbail@nvidia.com>
(cherry picked from commit 5fdb5ceb9135f7a1925a07003b2153f898231308)
Signed-off-by: Yun Zhou <yunz@nvidia.com>