SDN-2899 fix fwding issue when AIS-{Proxy|Target} are on the same node
Since the Host and OVN share the host MAC in shared gateway mode, the
reply packets from the Pod's public interface towards the Host's MAC is
getting forwarded to the Host instead of OVN.

In this case, we are hitting NORMAL action and OVS switch is forwarding
the packet to the Host.

Originally, the packet came from an overlay Pod in OVN, hit the L3 GW,
got forwarded to default gateway, got sent back to the host, then got
forwarded to the target Pod through the underlay link. The reply packet,
instead of heading back to OVN is heading back to host.

The fix here is to add a OpenFlow rule which steers the packet coming
from the localnet patch port with dst mac of Host to table=1 where we
check if the request packet originally came from the Host or OVN.

Signed-off-by: Girish Moodalbail <gmoodalbail@nvidia.com>
(cherry picked from commit 0a058e1ced8a3d1498a94eed9f823bc11937ec0c)
Signed-off-by: Yun Zhou <yunz@nvidia.com>