Fix e2e flakes in kubevirt
This MR includes fixes for known upstream flakes in the
KubeVirt E2E test. There are more changes upstream in this
test that are not yet backported, and changes that are
currently under review upstream.
See: https://github.com/ovn-org/ovn-kubernetes/pull/4755

Notably, some tests had to be disabled, as they were not
accounting for changes in OVN that are going to be merged.
See: https://github.com/ovn-org/ovn-kubernetes/pull/4457

This change is a stopgap until the upstream changes are
merged and backported. We need it to have a KubeVirt test
in the downstream CI/CD.
See: https://jirasw.nvidia.com/browse/SDN-2794

Signed-off-by: Flavio Fernandes <ffernandes@nvidia.com>