udn, nad template: Set explicit default MTU for locanet topology
For Localnet topology, the interface MTU should be default MTU (1500)
because the underlay is not part of the SDN and should not compensate
for the SDN overhead.

When MTU is unset, set the MTU explicitly in the NAD with default MTU (1500).

Signed-off-by: Or Mergi <ormergi@redhat.com>
(cherry picked from commit 87e93352bd06a776528368058ece87ff98f10421)
Signed-off-by: Yun Zhou <yunz@nvidia.com>