gateway_shared_intf: drop arps on bridgename when HNA is enabled.
For HNA (host network acceleration), we need to drop gARP packets sent by metallb
on the bridge's internal port.

Remove this commit after HNA is fleet wide and metallb skilift to add bridgeName
to metallb-excludel2 configmap is complete fleetwide in NGN

To make it more problem specific, add a comparision with "brbond0" and
exclude br2 nodes.

Signed-off-by: Guru Shetty <gurus@nvidia.com>