pbr: Add istio SNAT IP in LRP for host access
Istio in ambient mode SNATs packets to 169.254.7.127
when kubelet health checks the pods. The packet enters
OVN via ovn-k8s-mp0

When the pod responds, the packet will go to the
cluster router. We need a logical router policy
to send the packet back to the switch with
ovn-k8s-mp0 as the next hop

Signed-off-by: Guru Shetty <gurus@nvidia.com>